Add vdash syntactic sugar. (#383)
This PR adds syntactic sugar for type signatures, the "vdash" `|-`
syntax. It lets us factor out any type, resource, label, or stack base
that appears both as an input and output. The basic idea is that instead
of writing a type signature like this:

```
[ A B -- A C ]
```

We can write it like this instead:

```
[ A |- B -- C ]
```

Here `A` is considered a "context" for turning `B` into `C`. The context
`A` is both passed as input and returned as output. This is purely a
syntactic convenience, so it's not like the context is treated any
differently from other inputs or outputs, but it can be useful both in
reducing the amount of repetition, and for documenting the intention of
treating some input/output like a context.

The justificiation for adding this syntax is that we are quite often
working with a resource, or set of resources & labels, that are passed
as both input and output.

Some examples to illustrate how the syntactic sugar works:

* `[ Int |- Bool ]` is short for `[ Int -- Int Bool ]`
* `[ Int |- Bool -- Str ]` is short for `[ Int Bool -- Int Str ]`
* `[ Int Bool |- Str ]` is short for `[ Int Bool -- Int Bool Str ]`
* `[ +Foo |- Bar -- Baz ]` is short for `[ +Foo Bar -- +Foo Baz ]`
* `[ first:Str last:Str |- Str ]` is short for `[ first:Str last:Str --
first:Str last:Str Str ]`

The vdash syntax can also be used with higher-order words. The question
is whether the context comes before or after the higher-order params.
Both are allowed and they mean different things. Let's see.

* If the context comes before the parameters, as in `[ Z |- ( A -- B ) C
-- D ]`, then this the same as writing `[ ( Z A -- Z B ) Z C -- Z D ]`.
I.e. the context is passed to both the parameters and the word before
them.

* If the context comes after the parameters, as in `[ ( A -- B ) Z |- C
-- D ]` then this is the same as writing `[ ( A -- B ) Z C -- Z D ]`.
I.e. the context only affects what comes after it.

Additionally, since parameters are again just type signatures, they may
specify their own context, so you could write something like `[ ( Y |- A
-- B ) Z |- C -- D ]` and it's equivalent to `[ ( Y A -- Y B ) Z C -- Z
D ]`.

This should account for most of cases where this new syntax is useful.
The final twist / edge case to consider is where a stack type variable
appears in the domain or codomain when a context is also given. This is
handled carefully, so that the context gets placed after the stack type
variable but before the rest of the domain/codomain. For example, `[ A
|- *x B -- *y C ]` is the same as `[ *x A B -- *y A C ]`. It is an error
to have a stack type variable in both the context and the
domain/codomain, just as it is an error to have a stack type variable
appear in a stack type that already has a base. The main justification
for supporting this edge case is for when we want to build generic
higher-order words over a fixed resource context, e.g. how a lot of
`+Str` words work.

---

An unrelated change that got accidentally bundled with this PR:

* ldip was changed to only remove the top layer of each stack label by
default, so e.g. if you use `ldip { f }` when the stack has two types in
the `x` stack label (e.g. `[ x:Int x:Str ]`), only the top item is
dipped, not both, so `f` can access `x:Int`.