Implement automatic resource sliding. (#388)
This PR implements an automation for working with multiple resources:
Whenever a specific sequence of concrete resource types is required, the
elaborator will emit the necessary resource stack permutation to place
them in the required order. In other words, it will slide resources past
each other automatically. The elaborator only does this with concrete
resource types, not with generic or unknown resources, so it only emits
a resource stack permutation in cases that would have previously been
considered a type error.

The main effect of this change has is that we no longer need to use
`rswap` and `rdip` manually when working with multiple resources, the
compiler does it for us.

To take full advantage of this automation, name dispatching was changed
so that it will dispatch on any of the top resource types, not just the
very top one. This is a breaking change: some code that used to be
unambiguous is now ambiguous, so it won't compile anymore. You can fix
this by qualifying the methods with the specific concrete type, e.g.
instead of `rdrop` you say `+File.rdrop`.