Big table overhaul.  (#392)
This PR implements modern syntax for table declarations, allows tables &
fields to be associated with a resource (the "state owner" for the
table), and implements memoized default values for fields.

These changes were done in service of keeping the existing patterns of
table use but trying to bring them closer to the resource model.

The syntax for table declarations is similar to that for struct
declarations. Here is an example.

```
table +World |- Person {
   name: Str
   age: Int
   alive: Bool { drop True }
   ~parents: List(Person)
   --
   def greet! [ +World |- Person -- ] {
      print("Hello "; name ; "!";)
   }
   def John [ +World |- Person ] { 
      "John" >name
      20 >age
      Person.Make!
   }
}
```

This example shows a `Person` table that is "owned" by the `+World`
resource. In order to create `Person` table entries and access its table
fields, we need the `+World` resource. The `+World |-` can be left off
to have no state owner, but I think in the future "no state owner" is
going to just default to `+World` as the state owner.

The `Person` table has four fields: `name`, `age`, `alive`, and
`~parents`.

The `name`, `age`, and `alive` fields are "regular fields" which must be
accessed through helper methods:

- `name` to get and `name!` to set the `name` field. 
- `age` to get and `age!` to set the `age` field.
- `alive` to get and `alive!` to set the `alive` field.

These are similar to struct field access helper methods, though not all
the same helper methods are implemented for tables. But that would be
nice to have.

Additionally, the `alive` field has a default value which is calculated
when the field is first accessed if it was not already set. The default
value is given by a method with the type `[ +World |- Person -- Bool ]`,
in this case it has the implementation `{ drop True }`. Once calculated,
it is set as the field's value.

The `~parents` field is a "raw field", which means you access it
directly through the name `~parents` and this gives you a
`Mut(List(Person))`, i.e. a mutable reference to the field. These are
unsafe, so discouraged, but sometimes you have to use them.

A few more things of interest:

- Even with a "state owner" resource, the table is still global state.
In the future it would be nice to tie the table's state to its state
owner, as well as track the state owner's identity in the type system to
ensure you can only use the table with the same state owner. But this is
a future problem.

- Although they are no longer as useful, standalone `field` declarations
have also been revamped, they can take a state owner and default values:
`field(Owner? |- name, IndexType, ValueType, DefaultValue?)`

- Raw fields are fields that start with a tilde, e.g. `~parents` above.
Raw fields cannot have default values. But you can do a lot with `Mut`
values. (Before this PR all fields were raw fields.)

- Field names have been restricted somewhat. They must now start with a
lowercase letter (for regular fields) or a tilde (for raw fields). This
means regular field names are also valid label names.

- `T.Make!` is a new constructor for tables which takes all the
non-default non-raw fields as input. Standalone fields are not included
as inputs to `T.Make!`, only inline fields. (For various reasons they
can't be included.) The advantage of using `T.Make!` is it ensures that
all the fields that need to be set have values.