Use new table syntax in mirth compiler. (#393)
This PR is mostly to use the modern table syntax in the mirth compiler.
A few things I also changed about tables in the process:

- I changed the type of field defaults from `[ +StateOwner |- Table --
FieldValue ]` to `[ +StateOwner self:Table |- Value ]`. This means you
don't need to `drop` in order to provide a constant default value. This
PR includes a bridging definitions to account for this change (see
`field-self`), it will be removed in the next PR.

- I made `Table.=` one of the methods that tables come with by default.

- Some of the table methods (e.g. `Table.index`, `Table.Nil`, `Table.=`,
...) don't really require the state owner, so I removed the state owner
input.

- Deleted `Table.alloc!` and `Table.nil` in favor of `Table.Alloc!` and
`Table.Nil`.