Implement label lambda \x syntax. (#395)
Implements an experimental "label lambda" syntax. 

`\x` pushes to label `x`, and then pops `x` and drops it at the end of
the current expression. So,

```\x ...```

Is the same as,

```>x ... x> drop```

This is a common enough usage pattern for labels that it's worth looking
at ways to make it nicer.