Automatically co-dip extraneous values, resources, and labels when applying an operation against a fixed stack base. (#396)
This PR introduces an automation intended to make writing higher-order
words easier: whenever we apply an operation that has a fixed stack type
base, such as against a stack type variable, the elaborator will now
"co-dip" any extra stack items from the bottom of the stack, perform the
operation, and then put those items back at the bottom. This automation
only activates in situations that would previously have been type
errors.

Before this PR, when writing higher-order words, we would have to be
very careful about what went on the stack. E.g. if we want to implement
a higher-order function that applies a unary function `f` to two
elements, while maintaining access to the overall stack context, we
would had to write something like this:

```
def on2(f) [ (*a b -- *a c) *a b b -- *a c c ] {
   dip(f) swap dip(f) swap
}
```
Notice how we had to shuffle the top of the stack to the bottom in order
to apply `f` the second time, and then we had to undo this. Similarly,
we were discouraged from using stack labels or resources because we
would have to dip them away manually:
```
def foo(f) [ ( *a -- *b ) *a Int -- *b ] {
    \x ...  ldip(x, f) ...
}
def bar(f) [ (*a -- *b Str ) *a -- *b Str ] {
   Str( "f = " ; rdip(f) ; )
}
```

With the change in this PR, the compiler can hide anything that isn't
necessary. We can write the previous examples like this:
```
def on2(f) [ (*a b -- *a c) *a b b -- *a c c ] {
   dip(f) f
}

def foo(f) [ (*a -- *b) *a Int -- *b ] {
   \x ... f ...
}

def bar(f) [ (*a -- *b Str ) *a -- *b Str ] {
   Str( "f = " ; f ; )
}
```
In summary, the compiler now hides the bottom of the stack, including
extra labels and resources, when necessary.