Add Arrays. (#399)
Adds a primitive `Array(a)` type that stores a packed array of `a`
values. Adds a `std.array` module to expose `Array` methods as well as a
`+Array(a)` resource for building and/or working with arrays. Only
arrays with values are supported so far.

As part of this change I also made it possible for `def-type` type
definitions to take parameters, and also improved the representation of
data so that every type in C99 has a uniform representation (always uses
the same TAG).

Closes #384