Change arrays to always have a default value. (#402)
This PR makes a change to mirth's arrays that should make them more
usable (fewer edge cases) & safer (no panics, no sudden resizing). When
creating an array, you have to specify a "default value". The purpose of
this is twofold:

1. By specifying a value up front, you are giving enough information to
the runtime to construct an array of the correct size right away.
Previously, `prim-array-new` would assume the array value was a certain
size. By doing this we also open the possibility of giving (unsafe)
pointer access to array data, so they can be used in lower level APIs
(e.g. C library interop).

2. The default value is stored and used whenever you reading from an out
of bounds index. It is also used to expand the array automatically when
writing to an out of bounds index.

In theory, this makes arrays behave more like functions `Nat -> a`, that
are constant everywhere except on an initial segment. We still keep
track of length, so arrays are somewhere in between finite sequences and
infinite functions (with finite support).

The array literal idiom is now: 
```
Array( x1 ; x2 ; ... ; x3 ; xdefault )
```
Note the default value is not pushed with `;` it is left on the stack
when building an array. It's supposed to suggest that the "rest of the
array" is filled with the default value.

E.g. `Array( 10 ; 20 ; 30 ; 0 )` creates an array with three elements
`10`, `20`, `30`, and the default value `0`.