Drop Array fill value. (#404)
Go back to not having a fill value for arrays.

This means out of bound array access results in panic once again. You
can recreate the old functionality by checking the length yourself
and/or calling `Array.expand` / `+Array.expand!` manually beforehand.

The problem of not knowing how to allocate an array without having a
value ahead of time was solved by not allocating anything for an empty
array, I.e. use a null pointer internally to represent an empty array.
This PR makes it a panic if you attempt to access an empty array's base
pointer. But since you wouldn't be able to do anything safely with a
pointer to an empty array, this is fine actually.