Fix Array.expand reference counts and add +Debug resource. (#405)
This PR adds a +Debug resource which can be used to trace strings,
values, resources, & the stack to stderr, as well as obtain certain
internal information (value reference counts and identity -- i.e. the
value payload as a U64, which is often a pointer). This replaces the
`??` word, which is still available as `debug:trace-stack`.

This PR also fixes a bug in prim-array-expand which would add a
reference, and adds a regression test for it (usinge the new +Debug
resource).