Change requirements for transparent types. (#408)
This PR changes the requirement for transparent types. A transparent
type is a thin wrapper type, a `data` type with a single constructor and
a single input, or equivalently, a `struct` with a single field. In this
PR the requirement is strengthened so that the input must have a
concrete type.

This precludes a type like this from being transparent: 
```
struct Wrapper(a) { a }
```
where the input is a type variable. Instead this will be treated as a
struct type.

The reason to strengthen the requirement is so that transparent types
can have their own type tag, so we can associate dynamic behavior with
transparent types. If a single input type with a type variable were
transparent, we would have to store the type variable's type tag
somewhere, as well as the transparent type's tag whenever it is boxed,
and so we would either run out of space or lose type information. By
making it a struct, the type information for the type variable is stored
alongside the variable, in the struct.