Add Self and +Self type macros (#414)
This PR adds `Self` and `+Self` as type macros that refer to the type
constructor of the enclosing definition block. E.g.

```
struct Foo {
    -- 
    def foo [ Self -- Str ] { drop "foo" }
}
```

The `Self` refers to `Foo`.

For resource types, there is `+Self`.

Note that `Self` and `+Self` refer specifically to the type constructor,
so if there are type parameters you must pass them explicitly, e.g.
```
data List(a) {
    Nil
    Cons [ a List(a) ]
    --
    def map(f) [ (a -- b) Self(a) -- Self(b) ] { ... }
}
```
This is just a first implementation. Possible future improvements:
making `Self` infer missing type parameters from context, make sure
`Self` works in constructor/struct type signatures, using `Self` as a
namespace for qualification purposes.

Also it would be good to ultimately decide whether `Self` refers to the
tycon namespace of the current definition, or the type of the lexically
enclosing definition block. Currently it's the latter. This distinction
is only relevant for inline monkey patching style definitions, e.g. in
`def Int.foo [ ... Self ... ]` the `Self` doesn't refer to `Int` but to
whatever block this definition is in. I think the current approach is
fine, is more consistent, but someone might expect that `Self` to refer
to `Int` instead.